fix: resolve mago analyze type-safety findings in module source
Fix the type errors mago analyze reports against a full Magento install:

- CheckCommand: cast ini_get('memory_limit') (string|false declared
  string); use the PHP_VERSION constant instead of phpversion(); read the
  MySQL version via fetchOne('SELECT VERSION()') instead of from(null).
- Type the loose getPaths() return via inline @var (array<string,string>)
  in CompatibilityChecker and ThemePath.
- Cast mixed operands/arguments: theme loop index, hrtime() metrics,
  stat() mode, and the getOption() flags used in a boolean expression.
- InspectorHints: widen render()'s $dictionary param to match the parent
  TemplateEngineInterface; cast the BP constant to string.
- BlockCacheCollector: guard the untyped layout blocks with is_object()
  so method_exists() is safe and both analyzers narrow the type (drops
  the now-redundant @phpstan-ignore lines).
- IncompatibilityDetector: suppress the lone mago const-array-shape
  finding with a scoped @mago-expect pragma (the shape is correct, as
  PHPStan confirms).
- Drop redundant (string) casts and an obsolete @var docblock.

+++ src/Service/SymlinkCleaner.php
@@ -94,7 +94,7 @@ class SymlinkCleaner
             return false;
         }
 
-        return (($stat['mode'] ?? 0) & 0o12_0000) === 0o12_0000;
+        return ((int) ($stat['mode'] ?? 0) & 0o12_0000) === 0o12_0000;
     }
 
     /**
